auth-service/requirements.txt
SQLAlchemy~=2.0.31
uvicorn~=0.30.1
bcrypt~=4.1.3
PyJWT~=2.8.0
fastapi~=0.111.0
pydantic~=2.7.4
pydantic-settings~=2.3.4
psycopg2-binary~=2.9.9
PyYAML~=6.0.1